+++ src/app/armadura-componet/armadura-componet.component.ts
@@ -1,8 +1,7 @@
-import { DataServisService } from './../data-servis.service';
 import { Component, OnInit } from '@angular/core';
 import { Router } from '@angular/router';
-import { personaje } from '../personajes.model';
-import { ServicioService } from '../servicio.service';
+import { ServicioService } from '../services/servicio.service';
+import { Personaje } from '../models/personajes.model';
 
 @Component({
   selector: 'app-armadura-componet',
@@ -10,11 +9,10 @@ import { ServicioService } from '../servicio.service';
   styleUrls: ['./armadura-componet.component.css']
 })
 export class ArmaduraComponetComponent implements OnInit {
-  personaje:personaje[]=[];
-  constructor(private router:Router, private miServicio:ServicioService, private agregarServis:DataServisService){}
-  ngOnInit(): void {
-    
-  }
+  personaje:Personaje[]=[];
+  constructor(private router:Router, private miServicio:ServicioService){}
+  
+  ngOnInit(): void {}
 
   volverHome(){
     this.router.navigate(['']);
@@ -22,7 +20,7 @@ export class ArmaduraComponetComponent implements OnInit {
 
 
   agregar(){
-    let personajeEnviado= new personaje(this.cuadroNombre, this.cuadroApellido,this.cuadroPoder,this.cuadroEdad);
+    let personajeEnviado= new Personaje(this.cuadroNombre, this.cuadroApellido,this.cuadroPoder,this.cuadroEdad);
     this.miServicio.mustraMensaje("Nombre de personaje: " + personajeEnviado.nombre)
     this.personaje.push(personajeEnviado);
     this.router.navigate(['']);
@@ -35,7 +33,5 @@ export class ArmaduraComponetComponent implements OnInit {
   cuadroPoder:string="";
   cuadroEdad:number=0;
   
-  
-
 }
