+++ src/app/home-componet/home-componet.component.html
@@ -1,10 +1,4 @@
-
-  
-  
-  <br>
-  <br>
-  
-  <div class="container">
+  <div class="container mt-5">
        <div class="col my-4"><h1>{{titulo}}</h1></div>
