+++ src/app/services/personaje-servis.service.ts
@@ -1,0 +1,22 @@
+import { ServicioService } from './servicio.service';
+import { Injectable } from '@angular/core';
+import { Personaje } from "../models/personajes.model";
+
+@Injectable()
+export class  PersonajesService{
+
+constructor(private servicoVentana: ServicioService){}
+
+    personaje:Personaje[]=[
+        new Personaje("goblin slayer","no tiene","ninguno",21),
+        new Personaje("gordon","freman","traje de protecion",27),
+        new Personaje("angel","corral","desconocido",20),
+        new Personaje("alerbert","wesker","uroboros",30)
+    ];
+
+    agregarPersonajeService(personaje: Personaje){
+        this.servicoVentana.mustraMensaje("Personaje agregado: " + personaje.nombre)
+        this.personaje.push(personaje);
+        console.log("dasd")
+    }
+}
